+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Issue preventing symlinked directories from working with `checkpoint update --pending`
+- (internal) Naming of change target reason `uses_target` to `uses ` for clarity
 
 ## [3.5.6] - 2024-11-23